Big ol commit full of stuff! I learned Go!
gobyexample contains the gobyexample.com pages for learning go in the order they
were on the website. I have not finished *all* the examples, and I don't think
I intend too.

lilmdserver is my first little project that I want to use. It uses a third party
package as an example, and the build in HTTP library (or it will when I get around
to it). I'm hoping I get it done during this week, but I dunno :\